// client/src/App/Models/Basket.ts
export interface Basket {
    id: number
    buyerId: string
    items: BasketItem[];
    paymentIntentId?:string;
    clientSecret?:string;
  }
  
  export interface BasketItem {
    productId: number
    name: string
    price: number
    pictureUrl: string
    brand: string
    type: string
    quentity: number
  }
  